#1015 move grids to separate tab components
