+++ frontend/src/components/ModalConceptscheme.vue
@@ -8,27 +8,15 @@
     <template #body>
       <UTabs v-model="activeTab" color="neutral" variant="link" :items="tabs" class="w-full">
         <template #labels>
-          <UTable
-            class="flex-1 min-h-0 rounded-lg border border-default"
-            :data="labelsWithAddRow"
-            :columns="labelColumns"
-          />
+          <ModalConceptschemeTabLabels :data="labelsWithAddRow" @add="addLabel" />
         </template>
 
         <template #notes>
-          <UTable
-            class="flex-1 min-h-0 rounded-lg border border-default"
-            :data="notesWithAddRow"
-            :columns="noteColumns"
-          />
+          <ModalConceptschemeTabNotes :data="notesWithAddRow" @add="addNote" />
         </template>
 
         <template #sources>
-          <UTable
-            class="flex-1 min-h-0 rounded-lg border border-default"
-            :data="sourcesWithAddRow"
-            :columns="sourceColumns"
-          />
+          <ModalConceptschemeTabSources :data="sourcesWithAddRow" @add="addSource" />
         </template>
       </UTabs>
     </template>
@@ -41,17 +29,20 @@
   </UModal>
 </template>
 
+<script lang="ts">
+export type TableRow<T> = T & {
+  isAddRow?: boolean;
+};
+</script>
+
 <script setup lang="ts">
 import { useAdminUiStore } from '@stores/admin-ui';
 import { useConceptschemeStore } from '@stores/conceptscheme';
 import { storeToRefs } from 'pinia';
 import { useI18n } from 'vue-i18n';
-import type { TableColumn, TabsItem } from '@nuxt/ui';
-import { computed, h, ref, resolveComponent } from 'vue';
+import type { TabsItem } from '@nuxt/ui';
+import { computed, ref } from 'vue';
 import { type Label, type Note, type Source } from '@models/util';
-import DOMPurify from 'dompurify';
-
-const UButton = resolveComponent('UButton');
 
 const { t } = useI18n();
 
@@ -92,83 +83,13 @@ const addSource = () => {
   console.log('Add source clicked');
 };
 
-/* Tables */
-type TableRow<T> = T & {
-  isAddRow?: boolean;
-};
-
+/* Table data */
 const labelsWithAddRow = computed<TableRow<Label>[]>(() => [
   ...((selectedConceptscheme.value?.labels ?? []) as TableRow<Label>[]),
   {
     isAddRow: true,
   } as TableRow<Label>,
 ]);
-
-const labelColumns: TableColumn<TableRow<Label>>[] = [
-  {
-    accessorKey: 'label',
-    header: t('grid.columns.labels.label'),
-    cell: ({ row }) => {
-      if (row.original.isAddRow) {
-        return h(UButton, {
-          label: t('actions.add'),
-          icon: 'i-lucide-plus',
-          color: 'primary',
-          variant: 'outline',
-          size: 'xs',
-          class: 'cursor-pointer',
-          onClick: addLabel,
-        });
-      }
-
-      return row.original.label;
-    },
-    meta: {
-      class: {
-        th: 'w-full',
-        td: 'w-full',
-      },
-    },
-  },
-  {
-    accessorKey: 'language',
-    header: t('grid.columns.labels.language'),
-    cell: ({ row }) => (row.original.isAddRow ? '' : t('languages.' + row.original.language)),
-  },
-  {
-    accessorKey: 'type',
-    header: t('grid.columns.labels.type'),
-    cell: ({ row }) => (row.original.isAddRow ? '' : t('labelTypes.' + row.original.type)),
-  },
-  {
-    id: 'actions',
-    header: t('grid.columns.labels.actions'),
-    cell: ({ row }) => {
-      if (row.original.isAddRow) {
-        return '';
-      }
-
-      return h('div', { class: 'flex items-center gap-1' }, [
-        h(UButton, {
-          as: 'a',
-          href: '#',
-          label: t('grid.columns.actions.edit'),
-          icon: 'i-lucide-pencil',
-          color: 'primary',
-          variant: 'outline',
-          size: 'xs',
-        }),
-        h(UButton, {
-          icon: 'i-lucide-trash-2',
-          color: 'error',
-          variant: 'ghost',
-          size: 'xs',
-          'aria-label': t('grid.columns.actions.delete'),
-        }),
-      ]);
-    },
-  },
-];
 
 const notesWithAddRow = computed<TableRow<Note>[]>(() => [
   ...((selectedConceptscheme.value?.notes ?? []) as TableRow<Note>[]),
@@ -177,135 +98,10 @@ const notesWithAddRow = computed<TableRow<Note>[]>(() => [
   } as TableRow<Note>,
 ]);
 
-const noteColumns: TableColumn<TableRow<Note>>[] = [
-  {
-    accessorKey: 'note',
-    header: t('grid.columns.labels.note'),
-    cell: ({ row }) => {
-      if (row.original.isAddRow) {
-        return h(UButton, {
-          label: t('actions.add'),
-          icon: 'i-lucide-plus',
-          color: 'primary',
-          variant: 'outline',
-          size: 'xs',
-          class: 'cursor-pointer',
-          onClick: addNote,
-        });
-      }
-
-      return h('div', {
-        innerHTML: DOMPurify.sanitize(row.original.note, { USE_PROFILES: { html: true } }),
-      });
-    },
-    meta: {
-      class: {
-        th: 'w-full',
-        td: 'w-full',
-      },
-    },
-  },
-  {
-    accessorKey: 'language',
-    header: t('grid.columns.labels.language'),
-    cell: ({ row }) => (row.original.isAddRow ? '' : t('languages.' + row.original.language)),
-  },
-  {
-    accessorKey: 'type',
-    header: t('grid.columns.labels.type'),
-    cell: ({ row }) => (row.original.isAddRow ? '' : t('noteTypes.' + row.original.type)),
-  },
-  {
-    id: 'actions',
-    header: t('grid.columns.labels.actions'),
-    cell: ({ row }) => {
-      if (row.original.isAddRow) {
-        return '';
-      }
-
-      return h('div', { class: 'flex items-center gap-1' }, [
-        h(UButton, {
-          as: 'a',
-          href: '#',
-          label: t('grid.columns.actions.edit'),
-          icon: 'i-lucide-pencil',
-          color: 'primary',
-          variant: 'outline',
-          size: 'xs',
-        }),
-        h(UButton, {
-          icon: 'i-lucide-trash-2',
-          color: 'error',
-          variant: 'ghost',
-          size: 'xs',
-          'aria-label': t('grid.columns.actions.delete'),
-        }),
-      ]);
-    },
-  },
-];
-
 const sourcesWithAddRow = computed<TableRow<Source>[]>(() => [
   ...((selectedConceptscheme.value?.sources ?? []) as TableRow<Source>[]),
   {
     isAddRow: true,
   } as TableRow<Source>,
 ]);
-
-const sourceColumns: TableColumn<TableRow<Source>>[] = [
-  {
-    accessorKey: 'citation',
-    header: t('grid.columns.labels.source'),
-    cell: ({ row }) => {
-      if (row.original.isAddRow) {
-        return h(UButton, {
-          label: t('actions.add'),
-          icon: 'i-lucide-plus',
-          color: 'primary',
-          variant: 'outline',
-          size: 'xs',
-          class: 'cursor-pointer',
-          onClick: addSource,
-        });
-      }
-      return h('div', {
-        innerHTML: DOMPurify.sanitize(row.original.citation, { USE_PROFILES: { html: true } }),
-      });
-    },
-    meta: {
-      class: {
-        th: 'w-full',
-        td: 'w-full',
-      },
-    },
-  },
-  {
-    id: 'actions',
-    header: t('grid.columns.labels.actions'),
-    cell: ({ row }) => {
-      if (row.original.isAddRow) {
-        return '';
-      }
-
-      return h('div', { class: 'flex items-center gap-1' }, [
-        h(UButton, {
-          as: 'a',
-          href: '#',
-          label: t('grid.columns.actions.edit'),
-          icon: 'i-lucide-pencil',
-          color: 'primary',
-          variant: 'outline',
-          size: 'xs',
-        }),
-        h(UButton, {
-          icon: 'i-lucide-trash-2',
-          color: 'error',
-          variant: 'ghost',
-          size: 'xs',
-          'aria-label': t('grid.columns.actions.delete'),
-        }),
-      ]);
-    },
-  },
-];
 </script>
